refactor: Make value property in UniqueEntityID class readonly
feat: Add expiresIn option to JWT signOptions in auth.module.ts

feat: Add findAllByUserId method to CategoryRepository interface

refactor: Update GetAllCategoriesUseCase to use findAllByUserId

feat: Add user authentication to GetAllCategoriesController

feat: Add user authentication to GetCategoryController

refactor: Update CreateCategoryUseCase to use ICategoryProps interface

feat: Add user authentication to GetCategoryUseCase

feat: Add user authentication to DeleteCategoryUseCase

feat: Add findAllByUserId method to PrismaCategoryRepository

feat: Add user authentication to DeleteCategoryController

feat: Add user authentication to UpdateCategoryController

+++ src/domain/category/application/use-cases/delete-category.use-case.ts
@@ -1,11 +1,26 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CategoryRepository } from '../repositories/category-repository';
 
 @Injectable()
 export class DeleteCategoryUseCase {
   constructor(private readonly categoryRepository: CategoryRepository) {}
 
-  async execute(id: string): Promise<void> {
-    return await this.categoryRepository.delete(id);
+  async execute(id: string, userId: string): Promise<void> {
+    const category = await this.categoryRepository.findById(id);
+
+    if (!category) {
+      throw new NotFoundException('Category not found');
+    }
+
+    if (category.userId !== userId) {
+      throw new ForbiddenException('You can only delete your own categories');
+    }
+
+    // Exclui a categoria se todas as verificações forem passadas
+    await this.categoryRepository.delete(id);
   }
 }
